authorization and fix validate

+++ src/main/java/com/app/zware/Controllers/CategoryController.java
@@ -1,10 +1,13 @@
 package com.app.zware.Controllers;
 
 import com.app.zware.Entities.Category;
+import com.app.zware.Entities.User;
 import com.app.zware.Service.CategoryService;
 import java.util.List;
 
+import com.app.zware.Service.UserService;
 import com.app.zware.Validation.CategoryValidator;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -27,8 +30,12 @@ public class CategoryController {
   @Autowired
   CategoryValidator categoryValidator;
 
+  @Autowired
+  UserService userService;
+
   @GetMapping("")
   public ResponseEntity<?> index() {
+    //Authorization: All
     List<Category> categoryList = categoryService.getCategory();
     if (categoryList.isEmpty()) {
       return new ResponseEntity<>("Empty categoryList", HttpStatus.NOT_FOUND);
@@ -38,43 +45,66 @@ public class CategoryController {
   }
 
   @PostMapping("")
-  public ResponseEntity<?> store(@RequestBody Category category) {
+  public ResponseEntity<?> store(@RequestBody Category category, HttpServletRequest request) {
+    //Authorization: Only Admin
+    User user = userService.getRequestMaker(request);
+    if(!user.getRole().equals("admin")){
+      return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
+    }
+
+    //validation
     String mess = categoryValidator.checkPost(category);
     if(!mess.isEmpty()){
+      //error
       return new ResponseEntity<>(mess, HttpStatus.BAD_REQUEST);
     } else {
+      //approve
       return new ResponseEntity<>(categoryService.createCategory(category), HttpStatus.OK);
     }
   }
 
   @GetMapping("/{categoryId}")
   public ResponseEntity<?> show(@PathVariable("categoryId") Integer categoryId) {
+    //Authorization: all
+
+    //validation
       String message = categoryValidator.checkGet(categoryId);
       if(!message.isEmpty()){
+        //error
         return new ResponseEntity<>(message,HttpStatus.BAD_REQUEST);
       } else {
+        //approve
         Category category = categoryService.getCategoryById(categoryId);
         return new ResponseEntity<>(category, HttpStatus.OK);
       }
   }
 
   @DeleteMapping("/{categoryId}")
-  public ResponseEntity<?> destroy(@PathVariable("categoryId") Integer categoryId) {
+  public ResponseEntity<?> destroy(@PathVariable("categoryId") Integer categoryId, HttpServletRequest request) {
+    //Authorization: Only Admin
+    User user = userService.getRequestMaker(request);
+    if(!user.getRole().equals("admin")){
+      return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
+    }
+    //validation
     String message = categoryValidator.checkDelete(categoryId);
     if (!message.isEmpty()) {
+      //error
       return new ResponseEntity<>(message, HttpStatus.NOT_FOUND);
     } else {
+      //approve
       categoryService.deleteCategoryById(categoryId);
       return new ResponseEntity<>("Category has been deleted successfully", HttpStatus.OK);
     }
   }
 
   @PutMapping("/{categoryId}")
-  public ResponseEntity<?> update(@PathVariable Integer categoryId, @RequestBody Category request) {
-//    String message = categoryValidator.checkGet(categoryId);
-//    if (!message.isEmpty()) {
-//      return new ResponseEntity<>(message, HttpStatus.NOT_FOUND);
-//    } else {
+  public ResponseEntity<?> update(@PathVariable Integer categoryId, @RequestBody Category request, HttpServletRequest userRequest) {
+    //Authorization: Only Admin
+    User user = userService.getRequestMaker(userRequest);
+    if(!user.getRole().equals("admin")){
+      return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
+    }
 
     //merge infor
     Category updateCate = categoryService.merge(categoryId, request);
@@ -82,8 +112,10 @@ public class CategoryController {
     //check validate
       String mess = categoryValidator.checkPut( categoryId,updateCate);
       if(!mess.isEmpty()){
+        //error
         return new ResponseEntity<>(mess,HttpStatus.BAD_REQUEST);
       } else {
+        //approve
         categoryService.update( updateCate);
         return new ResponseEntity<>(updateCate, HttpStatus.OK);
       }
